requirements.txt
pandas
matplotlib
numpy
tensorflow==2.15.*
tensorflow-hub
scikit-learn
python-dotenv
datasets
transformers<=4.54.1
